Support jinja format for compare recipe
